requirements.txt
streamlit==1.32.0
playwright==1.42.0
python-dotenv==1.0.1
python-docx==1.1.0
chardet==5.2.0
pillow==10.2.0
beautifulsoup4==4.12.3
browser-use
python-docx
docxpython
tempfile
pypandoc
pyautogui
keyboard
language_tool_python
pyperclip
difflib

